requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
click>=8.0.0
markdown>=3.5.0
pygments>=2.16.0
pydantic>=2.0.0
requests>=2.31.0
rich>=13.0.0